requirements.txt
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
xgboost>=1.6
lightgbm>=3.3
joblib>=1.1
pytest>=7.0
requests>=2.28
requests-oauthlib>=1.3
tenacity>=8.0
python-dotenv>=1.0